Restructure file feature
