+++ src/main/kotlin/com/supa/spring/supaspring/file/domain/FileRepository.kt
@@ -1,0 +1,7 @@
+package com.supa.spring.supaspring.file.domain
+
+import com.supa.spring.supaspring.file.application.BucketItemDto
+
+interface FileRepository {
+    fun getFilesInBucket(bucketId: String): List<BucketItemDto>
+}
